update, refactor, seperate human, alpha-beta and board
