+++ C++_Player/common/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(common INTERFACE)
+
+target_include_directories(common INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+
+set_target_properties(common PROPERTIES CXX_STANDARD 14)
