+++ C++_Player/common/gameplayer.h
